src/main/python/community-detector/requirements.txt
numpy==1.14.2
python-igraph==0.7.1.post6
scipy==1.0.1
